enable constant propagation for union split signatures
This commit adopts the unmerged PR <https://github.com/JuliaLang/julia/pull/ #39305>,
and we will enable constant propagation for each union-split signature.

It should improve the analysis accuracy at the cost of possible
performance reduction (the nanosoldier benchmark should reveal it, but
currently the machine is down and the original PR is pending now because
of the reason).

Failures of constant propagation can sometimes lead to lots of false
positive error reports, which really spoils JET.jl's usability and
reliability. In particular, more correct constant analysis filtering
introduced in #110 reveals lots of hidden false positive errors
(e.g. see `@report_call rand(Bool)`), which are not easy to be "fixed".
I believe JET should eliminate them even if we have to wait more for an
analysis to finish, otherwise JET'd be an noisy and unusable analyzer.

Once the PR is merged, we can remove the hacky overloading of
`abstract_call_gf_by_type`, but the code can also be used as a new
"legacy code" for older versions of Julia.