re-fix #151
After #156, `fix_self_references!` replaces original self-referencing
`Symbol` with the un-canonicalized version of it, so that JET never uses
`Module` object in `CodeInfo` level.
That's especially because `Module` object is prohibited in module usage
expressions.

It turns out that the replaced `Symbol` can be something like
`var"Main.##JETVirtualModule##"`, and then abstract interpretation will
report false errors, just because it's not a valid global-ref.

This PR fixes it by only using symbolic replacements in module usages,
and keep to use `Module` object in other parts.