make `concretization_patterns` configuration work on surface-level AST, and fix the concretization of `@enum` macro expansion !
It turns out that it's so hard to setup a generalized statement
selection logic which works correctly on the expression expanded from
`@enum` macro; I'd like to omit the detail here, but it's fairly complex
and especially the [`let insts = (Any[ $(esc(typename))(v) for v in $values ]...,)`](https://github.com/JuliaLang/julia/blob/e5d7ef01b06f44cb75c871e54c81eb92eceae738/base/Enums.jl#L198)
part requires the logic to choose statements that only pushes elements
into a slot representing an array (`insts`), which is very hard to be
generalized.

So to fix this issue, I'd like to handle it as one of pre-defined
concretization patterns.
Since it works on macro, now `concretization_patterns` needs to work on
surface level AST, and the most of the changes in this commit comes
from that part. It's breaking but it should be more user-friendly if
the configuration works o surface level AST instead of lowered representation.