rework caching logic
The primary motivation for this is to eliminate as much dirty overloads
around `typeinf` as possible. Also I'm trying to avoid hijacking `InferenceResult.src`
field to keep JET's analysis result by maintaining a dictionary within
`AnalyzerState` to keep reports associating with `InferenceResult`
(which is essentially the same approach as what we are doing with Cthulhu).