`report_call` with a supremum matching method (#428)
fixes #427

`report_call`'s interface is designed to align with `code_typed` and the
family, but it expects a single matching method. That part is more like
`Base.which`. In this sense I guess we want to find a supremum matching
method as an entry signature, rather than complaining a given signature
is abstract or something.