virtualprocess: add special handling for Preferences.toml
This commit implements a hack to prevent Preferences.jl from throwing
errors for virtual modules. This is a very dirty solution for #497,
but ideally this should be replaced with a better solution if available.
Without this monkey patching, we will encounter something like:
```
│ ArgumentError: Module XXX does not correspond to a loaded package!
│ Stacktrace:
│  [1] get_uuid(m::Module)
│    @ Preferences ~/.julia/packages/Preferences/VmJXL/src/utils.jl:8
│  [2] var"@load_preference"(__source__::LineNumberNode, __module__::Module, key::Any, default::Any)
│    @ Preferences ~/.julia/packages/Preferences/VmJXL/src/Preferences.jl:45
```
We can't use the CassetteOverlay-like mechanism for a cleaner implementation,
since Preferences.jl might be called within `macroexpand` or `lower` of
the main `_virtual_process!` loop, where we don't have control over
execution. Note that this hack does not allow Preferences.jl to read or
set configurations correctly. A potential TODO is to read the
configurations of virtualized package being analyzed.