enable precompilation cache (#593)
With the introduction of `CodeInstance` serialization support in
JuliaLang/julia#52852, it's now reasonable to experiment with persisting
the analysis cache during precompilation.

There's an important consideration: the `codeinst.max_world` might
occasionally be set to `WORLD_AGE_REVALIDATION_SENTINEL` on the C side
under certain conditions. This assignment should ideally be reserved for
proper `CodeInstance`s (but it is not currently) and supposed to be
fixed up by a subsequent serialization pass. To address this, this
commit also adds an `__init__` hook that overrides such `max_world`s
at package loading time.