top-level: allow analysis entry point to be specified by name (#641)
When `analyze_from_definitions` is specified as `name::Symbol`, JET
starts its analysis using the interpreted method signature whose name is
equal to `name` as the analysis entry point. For example, when analyzing
a script that uses `@main` to specify the entry point, it would be
convenient to specify `analyze_from_definitions = :main`.