add `--trim` analyzer
This --trim specialized analyzer was previously managed by the JETLS
repository, but since it can be used without JETLS, it should be managed
as an independent analyzer.

While it's still uncertain whether this will be integrated into JET
proper, at least in the current experimental stage, it's more convenient
to make it usable by simply installing JET from this branch. Therefore,
this is being submitted as a draft PR for now.